[SOCIALGH-3] Use mixins for JSON mapping
This commit removes Jackson annotations from the domain objects and
replaces them with a number of mixins. This brings the GitHub binding
into line with the other bindings such as Facebook and Twitter.

As part of this change, properties have been made read-only where the
GitHub API does not allow the value to be changed.